+++ tests/unit/models/test_deployment.py
@@ -1,4 +1,7 @@
+import pytest
+
 from dbx.api.config_reader import ConfigReader
+from dbx.models.deployment import DeploymentConfig, EnvironmentDeploymentInfo
 from tests.unit.conftest import get_path_with_relation_to_current_file
 
 
@@ -16,3 +19,29 @@ def test_workflows_consistency(capsys):
     _new = ConfigReader(configs_path / "10-multitask-job-workflows.yaml").get_environment("default")
 
     assert _old == _new
+
+
+def test_non_unique_environments():
+    _env1 = EnvironmentDeploymentInfo(name="n1", payload={})
+    _dc = DeploymentConfig(environments=[_env1, _env1])
+    with pytest.raises(Exception):
+        _dc.get_environment("n1")
+
+
+def test_build_payload(capsys):
+    _payload = DeploymentConfig.prepare_build({"build": {"commands": ["sleep 5"]}})
+    res = capsys.readouterr()
+    assert "No build logic defined in the deployment file" not in res.out
+    assert _payload.commands is not None
+
+
+def test_build_payload_warning(capsys):
+    _payload = DeploymentConfig.prepare_build({})
+    res = capsys.readouterr()
+    assert "No build logic defined in the deployment file" in res.out
+
+
+def test_legacy_build_conflict():
+    with pytest.raises(ValueError) as exc_info:
+        DeploymentConfig.from_legacy_json_payload({"build": {"some": "value"}})
+    assert "Deployment file with a legacy syntax" in str(exc_info)